Move Planet and Asteroid into universe.py
